# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Forma única de todos los errores. La UI la traduce a un mensaje por `code`.",
        "properties": {
          "code": {
            "description": "Identificador estable del error, p. ej. NotFoundError",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Datos extra del error",
            "title": "Detail"
          },
          "message": {
            "description": "Mensaje en español, ya legible",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "SaveSecretIn": {
        "description": "Lo que entra. `value` es lo único que no vuelve a salir por ninguna ruta.",
        "properties": {
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "kind": {
            "$ref": "#/components/schemas/SecretKind"
          },
          "name": {
            "description": "Host, URL del sitio o proveedor",
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scope": {
            "default": "",
            "description": "Vacío = toda la organización. Si no, el requerimiento al que pertenece.",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "default": "",
            "description": "Usuario, cuando es un par usuario/contraseña",
            "title": "Username",
            "type": "string"
          },
          "value": {
            "description": "El secreto. No se devuelve nunca.",
            "minLength": 8,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name",
          "value"
        ],
        "title": "SaveSecretIn",
        "type": "object"
      },
      "SecretKind": {
        "description": "Para qué sirve el secreto. Determina quién lo pide y cómo se usa.",
        "enum": [
          "git_token",
          "site_credential",
          "llm_api_key",
          "azure_devops"
        ],
        "title": "SecretKind",
        "type": "string"
      },
      "SecretOut": {
        "description": "Un secreto, tal y como se puede enseñar.\n\n**No tiene campo para el valor, y es deliberado**: se construye desde `SecretMetadata`, que\ntampoco lo tiene, así que ninguna ruta puede devolverlo por descuido. `hint` son los cuatro\núltimos caracteres, lo justo para distinguir dos tokens sin reconstruir ninguno.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "hint": {
            "default": "",
            "title": "Hint",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "kindLabel": {
            "title": "Kindlabel",
            "type": "string"
          },
          "lastUsedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastusedat"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revokedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revokedat"
          },
          "revokedBy": {
            "default": "",
            "title": "Revokedby",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "username": {
            "default": "",
            "title": "Username",
            "type": "string"
          },
          "uses": {
            "default": 0,
            "title": "Uses",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "kindLabel",
          "name",
          "scope",
          "status",
          "createdBy",
          "createdAt",
          "updatedAt"
        ],
        "title": "SecretOut",
        "type": "object"
      },
      "SecretPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SecretOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "SecretPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend del orquestador multiagente. El dominio gira alrededor de un Requerimiento: adjuntos, plan de agentes, ejecuciones y entregables.",
    "title": "Orquestador de revisión de requerimientos",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/secrets": {
      "get": {
        "description": "Qué secretos hay configurados: tipo, a qué se aplican, quién y cuándo. **Sin valores.**",
        "operationId": "index_api_secrets_get",
        "parameters": [
          {
            "description": "Filtrar por ámbito. Ausente = todos; vacío = organización.",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por ámbito. Ausente = todos; vacío = organización.",
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Index",
        "tags": [
          "secretos"
        ]
      },
      "post": {
        "description": "Guarda o **rota** un secreto.\n\nGuardar otra vez el mismo tipo, nombre y ámbito sustituye el valor y levanta una revocación\nanterior: es la rotación. La respuesta son metadatos.",
        "operationId": "save_api_secrets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSecretIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Save",
        "tags": [
          "secretos"
        ]
      }
    },
    "/api/secrets/{secret_id}": {
      "delete": {
        "description": "Revoca un secreto: deja de funcionar de inmediato y su valor se borra de la base.",
        "operationId": "revoke_api_secrets__secret_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "secret_id",
            "required": true,
            "schema": {
              "title": "Secret Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Revoke",
        "tags": [
          "secretos"
        ]
      }
    },
...
  }
}